Translator: add check for reserved keywords.
This adds another Param field, Arg. This allows us to specify a
different name for the parameters.

This is needed to avoid reserved keyword collisions. If we have, for
example, a field with the name `type`, this would have previously
assigned the arguments as `func List(..., type string, ...)`. This
causes the Go compiler to fail.